JobPathEditor: show error when trying to open invalid jobId
Modify StopModel::loadJobStops() to return false for invalid jobs.
Clear JobPathEditor, disable it and show error to user.
This avoids confusing empty job which does not exist, but for which
JobPathEditor would ask user to save changes or delete because it has no
stops.
But the job does not exist so it cannot be removed nor it's changes
saved which would produce further error messages and really confuse the
user about what is happening.